backend/requirements.txt
fastapi[standard]
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic
python-jose==3.3.0
passlib==1.7.4
alembic==1.12.1
redis==5.0.1
email-validator==2.1.0
celery==5.3.4
sentry-sdk==1.32.0
prometheus-client==0.19.0
python-json-logger==2.0.7
FastAPI-SQLAlchemy
SQLAlchemy
psycopg2-binary==2.9.10
python-dotenv
pydantic-settings==2.6.0

# Development dependencies
black
pipenv
mypy-extensions
python-editor